requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
redis==5.1.1
httpx==0.27.2
anthropic==0.39.0
pypdf==5.0.1
python-docx==1.1.2
pyyaml==6.0.2
email-validator==2.2.0
